Refactor outputs state management
